Quantile loss function

A generalized version of the quantile loss. The model can predict multiple quantiles at once.

loss_quantile(...)

Arguments

quantiles

List of quantiles (numeric vector with values between 0 and 1).

Details

Loss value for a single sample-timestep-quantile is computed as: $$QL(y_t, \hat{y}_t, q) = max(q(y_t - \hat{y}_t), (q - 1)(y_t - \hat{y}_t))$$ or equivalently as : $$QL(y_t, \hat{y}_t, q) = max(q(y_t - \hat{y}_t), 0) + max((1 - q)(\hat{y}_t - y_t), 0)$$

When multiple quantiles are defined, the generalized, averaged loss is computed according to the equation: $$\mathcal{L}(\Omega, W) = \Sigma_{y_t \in \Omega}\Sigma_{q \in \mathcal{Q}}\Sigma^{\tau_{max}}_{\tau=1} = \frac{QL(y_t, \hat{y}(q, t - \tau, \tau), q)}{M_{\tau_{max}}}$$

The loss function is computed as below when reduction = 'auto or reduction = 'mean'.

References

Quantile loss function for machine learning

Temporal Fusion Transformer

Original TFT implementation by Google

Examples

y_pred <- array(runif(60), c(2, 10, 3))
y_true <- array(runif(20), c(2, 10, 1))

# As a callable object
loss_quantile(quantiles = c(0.1, 0.5, 0.9), reduction = 'auto')(y_true, y_pred)
#> tf.Tensor(0.1777453, shape=(), dtype=float32)
loss_quantile(quantiles = c(0.1, 0.5, 0.9), reduction = 'sum')(y_true, y_pred)
#> tf.Tensor(10.664718, shape=(), dtype=float32)

# As a function
loss_quantile(y_true, y_pred)
#> Error in py_call_impl(x, dots$args, dots$keywords): RuntimeError: Evaluation error: ValueError: Invalid Reduction Key: [[[0.69282902 0.07045285 0.39015057]
#>   [0.01800033 0.23995367 0.91973274]
#>   [0.20727386 0.48404959 0.17480934]
#>   [0.5237788  0.77049604 0.94080361]
#>   [0.52686886 0.68345829 0.49145884]
#>   [0.52256139 0.51826047 0.39596174]
#>   [0.28253671 0.58181476 0.32362742]
#>   [0.88777969 0.73196241 0.13862333]
#>   [0.01252248 0.68346003 0.54214726]
#>   [0.63986688 0.47415865 0.61564752]]
#> 
#>  [[0.44852693 0.17171853 0.4286302 ]
#>   [0.59769177 0.75747841 0.28352258]
#>   [0.35535566 0.03161492 0.91918637]
#>   [0.74245041 0.64168958 0.53368891]
#>   [0.59939877 0.96783543 0.05324609]
#>   [0.11337455 0.16169031 0.24713029]
#>   [0.64521732 0.13130664 0.9796318 ]
#>   [0.43927262 0.53507087 0.76529944]
#>   [0.81183512 0.59407257 0.49167133]
#>   [0.71063419 0.67760077 0.92482105]]]. Expected keys are "('auto', 'none', 'sum', 'sum_over_batch_size')"
#> .