Setup OxCal

Setup OxCal

Usage

oxcal_configure(
  command = NULL,
  os = NULL,
  ask = TRUE,
  install_location = NULL,
  install_url = NULL,
  verbose = getOption("aion.verbose")
)

Arguments

command

A character string specifying the path to the OxCal directory.

os

A character string specifying the operating system of the workstation. It must be one of "Linux", "Windows" or "Darwin". If NULL (the default), the operating system will be determined automatically (see Sys.info()).

ask

A logical scalar: if OxCal is not installed, should the user be asked to download it? If FALSE and R is being used interactively, will raise an error if the OxCal executable cannot be found. If FALSE and R is not being used interactively, will try to download OxCal.

install_location

A character string specifying the directory to extract OxCal files to. It will be created if necessary (see utils::unzip()).

install_url

A character string specifying the url from which OxCal could be installed.

verbose

A logical scalar: should status updates be displayed?

Value

Invisibly returns the path to the OxCal executable.

Details

Downloads the latest version of Oxcal (if needed) and sets the executable path correctly.

See also

Author

N. Frerebeau

Examples

if (FALSE) {
## Download OxCal
oxcal_configure()

## Custom script
scr <- 'Plot()
 {
  Sequence("Sequence1")
  {
   Boundary("Begin");
   Phase("Phase1")
   {
    R_Date("Lab-1",5000,25);
    R_Date("Lab-2",4900,37);
   };
   Boundary("Between");
   Phase("Phase2")
   {
    R_Date("Lab-3",4800,43);
   };
   Boundary("End");
  };
 };'
out <- oxcal_execute(scr)
res <- oxcal_parse(out)

plot(res)
plot(res, likelihood = TRUE, posterior = FALSE)
plot(res, likelihood = FALSE, posterior = TRUE)
}