Gated Linear Unit

In such form introduced in Language modeling with gated convolutional networks by Dauphin et al., when it was used in sequence processing tasks and compared with gating mechanism used in LSTM layers. In the context of time series processing explicitly proposed in Temporal Fusion Transformer.

layer_glu(object, units, activation = NULL, return_gate = FALSE, ...)

Arguments

object

What to compose the new Layer instance with. Typically a Sequential model or a Tensor (e.g., as returned by layer_input()). The return value depends on object. If object is:

missing or NULL, the Layer instance is returned.

a Sequential model, the model with an additional layer is returned.

a Tensor, the output tensor from layer_instance(object) is returned.

units

Positive integer, dimensionality of the output space.

activation

Name of activation function to use. If you don't specify anything, no activation is applied (ie. "linear" activation: a(x) = x).

return_gate

Logical - return gate values. Default: FALSE

Value

Tensor of shape batch_size, ..., units. Optionally, it can also return a weights tensor with identical shape.

Details

Computed according to the equation: $$GLU(\gamma) = \sigma(W\gamma + b) \odot (V\gamma + c)$$

References

Dauphin, Yann N., et al. (2017). Language modeling with gated convolutional networks.. International conference on machine learning. PMLR

Lim, Bryan et al. (2019). Temporal Fusion Transformers for Interpretable Multi-horizon Time Series Forecasting. arXiv

Implementation PyTorch

Implementation PyTorch

Examples

library(keras)

# ================================================================
#             SEQUENTIAL MODEL, NO GATE VALUES RETURNED
# ================================================================

model <-
  keras_model_sequential() %>%
  layer_glu(10, input_shape = 30)
#> Loaded Tensorflow version 2.8.0

model
#> Model: "sequential"
#> ________________________________________________________________________________
#>  Layer (type)                       Output Shape                    Param #     
#> ================================================================================
#>  glu (GLU)                          (None, 10)                      620         
#> ================================================================================
#> Total params: 620
#> Trainable params: 620
#> Non-trainable params: 0
#> ________________________________________________________________________________

output <- model(matrix(1, 32, 30))
dim(output)
#> [1] 32 10
output[1,]
#> tf.Tensor(
#> [-0.36247694 -0.5709695   0.24279653  0.5389923   0.35818368 -0.3389663
#>  -0.11138822 -0.54544526  0.29529864  0.09058831], shape=(10,), dtype=float32)

# ================================================================
#                     WITH GATE VALUES RETURNED
# ================================================================

inp  <- layer_input(30)
out  <- layer_glu(units = 10, return_gate = TRUE)(inp)

model <- keras_model(inp, out)

model
#> Model: "model"
#> ________________________________________________________________________________
#>  Layer (type)                       Output Shape                    Param #     
#> ================================================================================
#>  input_1 (InputLayer)               [(None, 30)]                    0           
#>  glu_1 (GLU)                        [(None, 10),                    620         
#>                                      (None, 10)]                                
#> ================================================================================
#> Total params: 620
#> Trainable params: 620
#> Non-trainable params: 0
#> ________________________________________________________________________________

c(values, gate) %<-% model(matrix(1, 32, 30))
dim(values)
#> [1] 32 10
dim(gate)
#> [1] 32 10

values[1,]
#> tf.Tensor(
#> [ 0.37070274 -0.50592345 -0.42771906 -0.25968853 -0.35275403  0.10566069
#>  -0.04980646 -0.27374062 -0.09268104  0.11610245], shape=(10,), dtype=float32)
gate[1,]
#> tf.Tensor(
#> [0.5550651  0.49924612 0.29802743 0.3920509  0.3225246  0.86447746
#>  0.6200736  0.6329947  0.15607809 0.42023954], shape=(10,), dtype=float32)