Quantile loss function

A generalized version of the quantile loss. The model can predict multiple quantiles at once.

Usage

loss_quantile(...)

loss_pinball(...)

Arguments

quantiles

List of quantiles (numeric vector with values between 0 and 1).

Details

Loss value for a single sample-timestep-quantile is computed as: $$QL(y_t, \hat{y}_t, q) = max(q(y_t - \hat{y}_t), (q - 1)(y_t - \hat{y}_t))$$ or equivalently as : $$QL(y_t, \hat{y}_t, q) = max(q(y_t - \hat{y}_t), 0) + max((1 - q)(\hat{y}_t - y_t), 0)$$

When multiple quantiles are defined, the generalized, averaged loss is computed according to the equation: $$\mathcal{L}(\Omega, W) = \Sigma_{y_t \in \Omega}\Sigma_{q \in \mathcal{Q}}\Sigma^{\tau_{max}}_{\tau=1} = \frac{QL(y_t, \hat{y}(q, t - \tau, \tau), q)}{M_{\tau_{max}}}$$

The loss function is computed as above when reduction = 'auto or reduction = 'mean'.

References

Quantile loss function for machine learning

Temporal Fusion Transformer

Original TFT implementation by Google

Examples

y_pred <- array(runif(60), c(2, 10, 3))
y_true <- array(runif(20), c(2, 10, 1))

# As a callable object
loss_quantile(quantiles = c(0.1, 0.5, 0.9), reduction = 'auto')(y_true, y_pred)
#> tf.Tensor(0.14314485, shape=(), dtype=float32)
loss_quantile(quantiles = c(0.1, 0.5, 0.9), reduction = 'sum')(y_true, y_pred)
#> tf.Tensor(8.588691, shape=(), dtype=float32)

# As a function
loss_quantile(y_true, y_pred)
#> Error in py_call_impl(x, dots$args, dots$keywords): RuntimeError: Evaluation error: ValueError: Invalid Reduction Key: [[[0.05870874 0.75366428 0.0465265 ]
#>   [0.53191079 0.81959078 0.57571952]
#>   [0.88330077 0.52788196 0.18498501]
#>   [0.68561851 0.0488573  0.6309801 ]
#>   [0.77633543 0.80887681 0.84715546]
#>   [0.16118917 0.94248154 0.1018559 ]
#>   [0.89832836 0.42212196 0.32075012]
#>   [0.35308511 0.74456039 0.08807754]
#>   [0.89902354 0.73170335 0.55505997]
#>   [0.98613548 0.58340203 0.67084564]]
#> 
#>  [[0.36135671 0.28129276 0.3444385 ]
#>   [0.62175414 0.62300187 0.56441954]
#>   [0.86254718 0.81325667 0.34436917]
#>   [0.86040729 0.34870888 0.04527512]
#>   [0.05345196 0.43990579 0.34601078]
#>   [0.47475089 0.7213679  0.2123746 ]
#>   [0.62241126 0.82508234 0.69795805]
#>   [0.04576325 0.89011892 0.52738856]
#>   [0.73725279 0.78378406 0.44278081]
#>   [0.59084461 0.99501221 0.68283359]]]. Expected keys are "('auto', 'none', 'sum', 'sum_over_batch_size')"
#> .